Add: protocol과 extension 관련 폴더에 protocol별로 파일 생성
